requirements.txt
Django==2.1.4
django-bootstrap-datepicker==1.3
django-bootstrap-datepicker-plus==3.0.4
django-bootstrap-datepicker-widget==0.1
django-bootstrap3==10.0.1
django-bootstrap4==0.0.6
django-crispy-forms==1.7.2
django-mssql==1.8
django-pyodbc-azure==2.1.0.0
django-python3-ldap==0.11.2
django-widget-tweaks==1.4.3
ldap3==2.5
pyasn1==0.4.4
pyodbc==4.0.23
pytz==2018.4
